All changes

+++ InfinApp.Web/Repositories/IActividadColaboradorRepository.cs
@@ -5,9 +5,9 @@ namespace InfinApp.Web.Repositories
     public interface IActividadColaboradorRepository
     {
         Task<List<ActividadColaborador>> ObtenerTodos();
-        Task<ActividadColaborador?> ObtenerPorId(int id);
+        Task<ActividadColaborador?> ObtenerPorId(long id);
         Task<int> Crear(ActividadColaborador model);
         Task<bool> Actualizar(ActividadColaborador model);
-        Task<bool> Eliminar(int id);
+        Task<bool> Eliminar(long id);
     }
 }
